requirements.txt
setuptools==57.5.0
celery
Django==4.1.7
django-celery-beat
djangorestframework==3.14.0
psycopg2-binary==2.9.3
pytz==2022.7.1
redis==4.5.1
kafka-python
drf-yasg
django-nose
coverage
